work_items: fix work item type not found exception
The field workItemType is now a String instead
of an enum, so that all possible values are accepted.
The available work item types for each project
are downloaded and cached when entering
the work items page.

Known limitations: some work item types don't work
in create/edit because they require some fields
which are not present in UI.