pull_requests: handle mention in comments
This is done by getting the user ids of all
the mentioned user, then replacing the ids
with the html mention tag (like in work items
mentions). When adding/editing a comment,
we have to replace back the html tag with
the markdown mention which is accepted
by azure api.