pull_requests: add/edit/delete comment in file
This commit adds the possibility to add/edit/delete a
comment to a specific line of code in a pull request file diff.

The logic to replace work items link and mentions
inside comments has been extracted
in `PullRequestHelper` mixin because it was duplicate
between pull request detail page and file diff page.

The UI code that shows a pull request comment has been
extracted in `PullRequestCommentCard` because it was duplicate in
pull request and file diff page.

Also, comments are now editable/deletable only by
the comment author.